requirements.txt
Flask
Flask-SQLAlchemy
oracledb
python-dotenv
werkzeug